Break the components as smaller can Resolved commented issue
